+++ packages/drawer/src/views/DrawerView.tsx
@@ -256,7 +256,7 @@ function DrawerViewBase({
             headerShown,
             headerStatusBarHeight,
             headerTransparent,
-            sceneContainerStyle,
+            sceneStyle,
           } = descriptor.options;
 
           return (
@@ -281,7 +281,7 @@ function DrawerViewBase({
                     descriptor.navigation as DrawerNavigationProp<ParamListBase>,
                   options: descriptor.options,
                 })}
-                style={sceneContainerStyle}
+                style={sceneStyle}
               >
                 {descriptor.render()}
               </Screen>