+++ packages/devtools/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0-rc.16](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-rc.15...@react-navigation/devtools@7.0.0-rc.16) (2024-10-27)
+
+### Features
+
+* add a logger devtool ([#12208](https://github.com/react-navigation/react-navigation/issues/12208)) ([8497f78](https://github.com/react-navigation/react-navigation/commit/8497f78858b6a8d8a6b397ccd441d46b0c8ed8c7)) - by @satya164
+
 # [7.0.0-rc.15](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-rc.14...@react-navigation/devtools@7.0.0-rc.15) (2024-10-11)
 
 **Note:** Version bump only for package @react-navigation/devtools